# GO-Backend/docs/swagger.yaml
basePath: /api
definitions:
  model.PhoneNumber:
    properties:
      id:
        type: string
      number:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  model.User:
    properties:
      address:
        type: string
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
      nic:
        type: string
      phones:
        items:
          $ref: '#/definitions/model.PhoneNumber'
        type: array
      photo:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple user management API using Fiber and MongoDB.
  title: Go Fiber User API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - multipart/form-data
      description: Update an existing user with optional photo upload (multipart/form-data)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User's full name
        in: formData
        name: name
        type: string
      - description: User's email address
        in: formData
        name: email
        type: string
      - description: National ID number
        in: formData
        name: nic
        type: string
      - description: User's address
        in: formData
        name: address
        type: string
      - description: Birthday in ISO format (e.g. 2006-01-02T15:04:05.000Z)
        in: formData
        name: birthday
        type: string
      - description: Gender (e.g. Male or Female)
        in: formData
        name: gender
        type: string
      - description: User's profile image (jpg/png)
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - Users
  /users/{id}/phones:
    get:
      consumes:
      - application/json
      description: Retrieve all phone numbers associated with a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PhoneNumber'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all phone numbers for a user
      tags:
      - Phones
    post:
      consumes:
      - application/json
      description: Add a new phone number to an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone number data
        in: body
        name: phone
        required: true
        schema:
          $ref: '#/definitions/model.PhoneNumber'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PhoneNumber'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new phone number for a user
      tags:
      - Phones
  /users/{id}/phones/{phoneId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific phone number for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phoneId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a phone number
      tags:
      - Phones
    put:
      consumes:
      - application/json
      description: Update an existing phone number for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phoneId
        required: true
        type: string
      - description: Updated phone data
        in: body
        name: phone
        required: true
        schema:
          $ref: '#/definitions/model.PhoneNumber'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PhoneNumber'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a phone number
      tags:
      - Phones
  /users/{id}/with-phones:
    get:
      consumes:
      - application/json
      description: Retrieve a specific user with all their phone numbers
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user with phone numbers
      tags:
      - Users
  /users/with-phones:
    get:
      consumes:
      - application/json
      description: Retrieve all users along with their associated phone numbers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users with their phone numbers
      tags:
      - Users
swagger: "2.0"